requirements.txt
transformers
datasets
accelerate
bitsandbytes
peft
hf-transfer
trl
hydra-core==1.3.2
matplotlib
wandb
tensorboard
fire
flash-attn
packaging
